Restructured the code into a more sensible structure, and made the linter happy
